replace util with validator in project groups
